+++ CodingameFramework/src/main/java/fr/code/variable/Constant.java
@@ -2,6 +2,6 @@ package fr.code.variable;
 
 public class Constant {
 
-  public static byte PLAYER_ID;
-  public static byte OPPONENT_ID;
+  public static int PLAYER_ID;
+  public static int OPPONENT_ID;
 }
